requirements.txt
tensorflow==2.4.0
tensorflow-gpu==2.4.0
keras==2.4.0
numpy==1.19.2
matplotlib<=3.2
pandas
scikit-image
Pillow
opencv-python
labelme
imutils